Ability to choose allometric calc method
Maestra/Maespa uses a quasi-geometric method of calculating masses from
tree height and diameter. Code was added to allow the user to use a
calculation method more in line with allometric and metabolic theory
(West, Brown, Enquist) in that the equation used is in the form
classically described by Huxley and Tessier.